rename overview.middleware to lpa-overview.middleware, some JSDoc
